requirements.txt
pandas
PyMySql
SQLAlchemy
streamlit-aggrid
python-dotenv
Babel